+++ spec/features/visitors/view_forecasts_spec.rb
@@ -1,58 +1,144 @@
 # frozen_string_literal: true
 
-# Feature: Forecasts page
-#   So that I can understand potential health dangers in my area
-#   As a visitor who has chosen a particular zone
-#   I want to see warnings and alerts for both pollutants and non-pollutants for 3 days
 RSpec.feature "Forecasts page", feature: true do
-  include ForecastSteps
+  include Features::ForecastHelper
 
-  scenario "See forecasts for 3 days" do
-    given_a_forecast_for_today
-    and_a_forecast_for_tomorrow
-    and_a_forecast_for_the_day_after_tomorrow
-    and_the_response_from_cercs_api_is_stubbed_accordingly
+  describe "Viewing forecasts" do
+    describe "Viewing forecasts for 3 days" do
+      before do
+        forecasts = [
+          # Given the forecasts for today, tomorrow and the day after tomorrow
+          Fixtures::API.zone_forecast(
+            day: :today,
+            air_pollution_status: :high,
+            pollen: :low,
+            temperature: :cold,
+            uv: :low
+          ),
+          Fixtures::API.zone_forecast(
+            day: :tomorrow,
+            air_pollution_status: :moderate,
+            pollen: :moderate,
+            temperature: :normal,
+            uv: :moderate
+          ),
+          Fixtures::API.zone_forecast(
+            day: :day_after_tomorrow,
+            air_pollution_status: :very_high,
+            pollen: :high,
+            temperature: :hot,
+            uv: :high
+          )
+        ]
+        stub_cerc_api_with(forecasts)
+      end
 
-    visit root_path
-    when_i_select_view_forecasts
-    then_i_see_the_forecasts_page
+      it "shows the forecasts for today, tomorrow and the day after tomorrow" do
+        ##
+        # When I select view forecasts
+        #
+        # Then I see the forecasts page
+        # and I see that the today tab is active
+        # and I see predicted air pollution status for each day
+        # and I see predicted uv level
+        # and I see predicted pollen level
+        # and I see predicted temperature level
+        ##
 
-    and_i_see_that_the_today_tab_is_active
-    and_i_see_predicted_air_pollution_status_for_each_day
-    and_i_see_predicted_uv_level
-    and_i_see_predicted_pollen_level
-    and_i_see_predicted_temperature_level
-  end
+        view_forecasts
 
-  scenario "See detail for tomorrow", js: true do
-    given_a_forecast_for_today
-    and_a_forecast_for_tomorrow
-    and_a_forecast_for_the_day_after_tomorrow
-    and_the_response_from_cercs_api_is_stubbed_accordingly
+        # See the forecasts page
+        expect(page).to have_content("Air quality forecast")
 
-    visit root_path
-    when_i_select_view_forecasts
-    and_i_switch_to_the_tab_for_tomorrow
+        # Today tab is active
+        expect(page).to have_css(".tab.today.active")
 
-    then_i_see_that_the_tomorrow_tab_is_active
-    and_i_see_predicted_uv_level_for_tomorrow
-    and_i_see_predicted_pollen_level_for_tomorrow
-    and_i_see_predicted_temperature_level_for_tomorrow
-  end
+        expect(page).to have_css(".tab.tomorrow.inactive")
+        expect(page).to have_css(".tab.day_after_tomorrow.inactive")
 
-  scenario "See detail for day after tomorrow", js: true do
-    given_a_forecast_for_today
-    and_a_forecast_for_tomorrow
-    and_a_forecast_for_the_day_after_tomorrow
-    and_the_response_from_cercs_api_is_stubbed_accordingly
+        expect(page).not_to have_css(".tab.tomorrow.active")
+        expect(page).not_to have_css(".tab.day_after_tomorrow.active")
 
-    visit root_path
-    when_i_select_view_forecasts
+        # Air pollution status for each day
+        expect_air_pollution_prediction(day: :today, value: :high)
+        expect_air_pollution_prediction(day: :tomorrow, value: :moderate)
+        expect_air_pollution_prediction(day: :day_after_tomorrow, value: :very_high)
 
-    and_i_switch_to_the_tab_for_day_after_tomorrow
-    then_i_see_that_the_day_after_tomorrow_tab_is_active
-    and_i_see_predicted_uv_level_for_day_after_tomorrow
-    and_i_see_predicted_pollen_level_for_day_after_tomorrow
-    and_i_see_predicted_temperature_level_for_day_after_tomorrow
+        # UV level
+        expect_prediction(category: :"ultraviolet-rays-uv", level: :low)
+
+        # Pollen level
+        expect_prediction(category: :pollen, level: :low)
+
+        # Temperature level
+        expect_prediction(category: :temperature, level: :low)
+      end
+
+      it "shows details for tomorrow", js: true do
+        ##
+        # When I select view forecasts
+        # And I switch to the tab for tomorrow
+        #
+        # Then I see that the tomorrow tab is active
+        # And I see predicted UV level for the day after tomorrow
+        # And I see predicted pollen level for tomorrow
+        # And I see predicted temperature level for tomorrow
+        ##
+
+        view_forecasts
+        switch_to_tab_for(:tomorrow)
+
+        # See that the tomorrow tab is active
+        expect(page).to have_css(".tab.tomorrow.active")
+
+        expect(page).to have_css(".tab.today.inactive")
+        expect(page).to have_css(".tab.day_after_tomorrow.inactive")
+
+        expect(page).not_to have_css(".tab.today.active")
+        expect(page).not_to have_css(".tab.day_after_tomorrow.active")
+
+        # Predicted UV level for tomorrow
+        expect_prediction(category: :"ultraviolet-rays-uv", level: :moderate)
+
+        # Predicted pollen level for tomorrow
+        expect_prediction(category: :pollen, level: :moderate)
+
+        # Predicted temperature level for tomorrow
+        expect_prediction(category: :temperature, level: :moderate)
+      end
+
+      it "shows details for the day after tomorrow", js: true do
+        ##
+        # When I select view forecasts
+        # And I switch to the tab for the day after tomorrow
+        #
+        # Then I see that the day after tomorrow tab is active
+        # And I see predicted UV level for the day after tomorrow
+        # And I see predicted pollen level for the day after tomorrow
+        # And I see predicted temperature level for the day after tomorrow
+        ##
+
+        view_forecasts
+        switch_to_tab_for(:day_after_tomorrow)
+
+        # See that the day after tomorrow tab is active
+        expect(page).to have_css(".tab.day_after_tomorrow.active")
+
+        expect(page).to have_css(".tab.today.inactive")
+        expect(page).to have_css(".tab.tomorrow.inactive")
+
+        expect(page).not_to have_css(".tab.today.active")
+        expect(page).not_to have_css(".tab.tomorrow.active")
+
+        # Predicted UV level for the day after tomorrow
+        expect_prediction(category: :"ultraviolet-rays-uv", level: :high)
+
+        # Predicted pollen level for the day after tomorrow
+        expect_prediction(category: :pollen, level: :high)
+
+        # Predicted temperature level for the day after tomorrow
+        expect_prediction(category: :temperature, level: :high)
+      end
+    end
   end
 end
